+++ src/modules/dish/dish.controller.spec.ts
@@ -73,11 +73,10 @@ describe('DishController', () => {
         it('should return detailed dish with translated ingredients', async () => {
             const id = '123';
             const lang: Language = 'en';
-            const dish = { description: 'description', ingredients: ['ingredient1', 'ingredient2'], recipeSections: [] } as any;
+            const dish = { description: 'description', ingredients: ['ingredient1', 'ingredient2'] } as any;
             const translatedDescription = 'translated description' as string;
             const translatedIngredients = ['translated1', 'translated2'] as any;
-            const translatedRecipe = ['translated1', 'translated2'] as any;
-            const expectedResult = { dish, description: translatedDescription, ingredients: translatedIngredients, recipe: translatedRecipe };
+            const expectedResult = { dish, description: translatedDescription, ingredients: translatedIngredients };
 
             jest.spyOn(dishService, 'getDishDetails').mockResolvedValue(dish);
             jest.spyOn(translationService, 'translateDish').mockResolvedValue(expectedResult);